Fix English spelling and grammar errors on demo page
The original page had a lot of mistakes in the English text that had been written, so I have forked this project and re-written this page's text so that it is more clear and easier to read. I hope that in doing so both English and non-English speakers will be able to understand eModal better.